Added user class with service and model
User model is the object class.
User service will either create a new user and push it into the database or
get info to make sure username and password are correct for login using orderByChild
and equalTo.
Signup has newUser, saveUser, and onSubmit to add the user
Login has onSubmit that calls the checkCredentials
Both are having issues